configure authentication

+++ SimbirHealth.Account/Models/Info/JwtInfo.cs
@@ -3,7 +3,7 @@
     public class JwtInfo
     {
         public JwtInfo() { }
-        public JwtInfo(string issuerName, string secretKey, double accessLiveHours, double accessLiveMinutes, double refreshLiveDays)
+        public JwtInfo(string issuerName, string secretKey, int accessLiveHours, int accessLiveMinutes, int refreshLiveDays)
         {
             IssuerName = issuerName;
             SecretKey = secretKey;
@@ -15,8 +15,8 @@
         public static readonly string SectionName = "JwtAuthenticationInfo";
         public string IssuerName { get; set; }
         public string SecretKey { get; set; }
-        public double AccessLiveHours { get; set; }
-        public double AccessLiveMinutes { get; set; }
-        public double RefreshLiveDays { get; set; }
+        public int AccessLiveHours { get; set; }
+        public int AccessLiveMinutes { get; set; }
+        public int RefreshLiveDays { get; set; }
     }
 }
